Make the routes more testable
 * Split GP_Router and GP_Route into different files
 * Do not exit in can_or_redirect() type methods. Rename them to cannot_and_redirect() type and return in the route if they have redirected or died

+++ gp-includes/routes/project.php
@@ -21,7 +21,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function personal_options_post( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$user = GP::$user->current();
 		$source_url_templates = $user->get_meta( 'source_url_templates' );
 		if ( !is_array( $source_url_templates ) ) $source_url_templates = array();
@@ -36,7 +36,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function import_originals_get( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$kind = 'originals';
 		gp_tmpl_load( 'project-import', get_defined_vars() );
 	}
@@ -44,7 +44,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function import_originals_post( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 
 		$format = gp_array_get( GP::$formats, gp_post( 'format', 'po' ), null );
 		if ( !$format ) {
@@ -73,16 +73,16 @@ class GP_Route_Project extends GP_Route_Main {
 	function edit_get( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		gp_tmpl_load( 'project-edit', get_defined_vars() );
 	}
 	
 	function edit_post( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$updated_project = new GP_Project( gp_post( 'project' ) );
-		$this->validate_or_redirect( $updated_project, gp_url_project( $project, '-edit' ) );
+		if ( $this->invalid_and_redirect( $updated_project, gp_url_project( $project, '-edit' ) ) ) return;
 		// TODO: add id check as a validation rule
 		if ( $project->id == $updated_project->parent_project_id )
 			$this->errors[] = __('The project cannot be parent of itself!');
@@ -101,7 +101,7 @@ class GP_Route_Project extends GP_Route_Main {
 		// TODO: just deactivate, do not actually delete
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		if ( $project->delete() )
 			$this->notices[] = __('The project was deleted.');
 		else
@@ -113,16 +113,16 @@ class GP_Route_Project extends GP_Route_Main {
 	function new_get() {
 		$project = new GP_Project();
 		$project->parent_project_id = gp_get( 'parent_project_id', null );
-		$this->can_or_redirect( 'write', 'project', $project->parent_project_id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->parent_project_id ) ) return;
 		gp_tmpl_load( 'project-new', get_defined_vars() );
 	}
 	
 	function new_post() {
 		$post = gp_post( 'project' );
 		$parent_project_id = gp_array_get( $post, 'parent_project_id', null );
-		$this->can_or_redirect( 'write', 'project', $parent_project_id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $parent_project_id ) ) return;
 		$new_project = new GP_Project( $post );
-		$this->validate_or_redirect( $new_project );
+		if ( $this->invalid_and_redirect( $new_project ) ) return;
 		$project = GP::$project->create_and_select( $new_project );
 		if ( !$project ) {
 			$project = new GP_Project();
@@ -137,7 +137,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function permissions_get( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$path_to_root = array_slice( $project->path_to_root(), 1 );
 		$permissions = GP::$validator_permission->by_project_id( $project->id );
 		$cmp_fn = lambda( '$x, $y', 'strcmp($x->locale_slug, $y->locale_slug);' );
@@ -161,7 +161,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function permissions_post( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		if ( 'add-validator' == gp_post( 'action' ) ) {
 			$user = GP::$user->by_login( gp_post( 'user_login' ) );
 			if ( !$user ) {
@@ -175,7 +175,7 @@ class GP_Route_Project extends GP_Route_Main {
 				'locale_slug' => gp_post( 'locale' ),
 				'set_slug' => gp_post( 'set-slug' ),
 			) );
-			$this->validate_or_redirect( $new_permission, gp_url_current() );
+			if ( $this->invalid_and_redirect( $new_permission, gp_url_current() ) ) return;
 			$permission = GP::$validator_permission->create( $new_permission );
 			$permission?
 				$this->notices[] = __('Validator was added.') : $this->errors[] = __('Error in adding validator.');
@@ -186,7 +186,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function permissions_delete( $project_path, $permission_id ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$permission = GP::$permission->get( $permission_id );
 		if ( $permission ) {
 			if ( $permission->delete() ) {
@@ -203,14 +203,14 @@ class GP_Route_Project extends GP_Route_Main {
 	function mass_create_sets_get( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$this->tmpl( 'project-mass-create-sets', get_defined_vars() );
 	}
 
 	function mass_create_sets_post( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$other_project = GP::$project->get( gp_post( 'project_id' ) );
 		if ( !$other_project ) {
 			$this->die_with_error( __('Project wasn&#8217;found') );
@@ -234,7 +234,7 @@ class GP_Route_Project extends GP_Route_Main {
 	function mass_create_sets_preview_post( $project_path ) {
 		$project = GP::$project->by_path( $project_path );
 		if ( !$project ) gp_tmpl_404();
-		$this->can_or_redirect( 'write', 'project', $project->id );
+		if ( $this->cannot_and_redirect( 'write', 'project', $project->id ) ) return;
 		$other_project = GP::$project->get( gp_post( 'project_id' ) );
 		if ( !$other_project ) {
 			$this->die_with_error( __('Project wasn&#8217;found') );
